Update the files for change the file names
