+++ src/components/BookList.tsx
@@ -1,16 +1,15 @@
 import  {  useEffect } from 'react';
-import './book-list.css'
+import '../styles/book-list.css'
 
-import SearchBar from './search';
-import { useCart } from './cartContext';
+import SearchBar from './SearchComponent';
+import { useCart } from '../context/CartContextToHandleCaer';
 
 const BookData = () => {
   const {cards,setCards}=useCart()
   const fetchData = async() => {
-    let a = await fetch('  https://68ca8400430c4476c349caf6.mockapi.io/books/Books')
-    let data = await a.json()
+    let fetchedData = await fetch('  https://68ca8400430c4476c349caf6.mockapi.io/books/Books')
+    let data = await fetchedData.json()
     setCards(data)
-    console.log(data)
    }            
    useEffect (()=> {
     fetchData()
