+++ src/types/AddToCartProps.ts
@@ -1,0 +1,8 @@
+export type cardItemType = {
+  id: number;
+  bookName: string;
+  price: number;
+  author: string;
+  bookImage: string;
+  quantity: number;
+}
