+++ src/types/BookCardProps.ts
@@ -1,0 +1,10 @@
+export type CardProps = {
+  id: number;
+  bookName: string;
+  price: number;
+  author: string;
+  bookImage: string;
+};
+
+
+
