+++ src/types/DetailsPage.ts
@@ -1,0 +1,9 @@
+export type Details={
+  id: number;
+  bookName: string;
+  price: number;
+  author: string;
+  bookImage: string;
+  quantity: number;
+  description:string;
+}
