+++ src/types/UserDetails.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id: string;
+  name: string;
+  email: string;
+  password: string;
+}
