--- src/types/prop-type.tsx
@@ -1,26 +1,0 @@
-export type CardProps = {
-  id: number;
-  bookName: string;
-  price: number;
-  author: string;
-  bookImage: string;
-};
-
-export type cardItemType = {
-  id: number;
-  bookName: string;
-  price: number;
-  author: string;
-  bookImage: string;
-  quantity: number;
-  
-}
-export type Details={
-  id: number;
-  bookName: string;
-  price: number;
-  author: string;
-  bookImage: string;
-  quantity: number;
-  description:string;
-}
